safari-600.7-branch/LayoutTests/js/dom/Promise-catch-in-workers-expected.txt

[Worker] Test Promise.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: ../resources/Promise-catch-in-workers.js
PASS [Worker] thisInInit === firstPromise is false
PASS [Worker] thisInInit === global is true
PASS [Worker] firstPromise instanceof Promise is true
PASS [Worker] secondPromise instanceof Promise is true
PASS [Worker] thisInFulfillCallback === firstPromise is false
PASS [Worker] thisInFulfillCallback === secondPromise is false
PASS [Worker] thisInFulfillCallback === global is true
PASS [Worker] result is "hello"
PASS [Worker] result is "bye"
PASS [Worker] fulfilled
PASS successfullyParsed is true

TEST COMPLETE
